+++ clease/montecarlo/base.py
@@ -15,14 +15,12 @@ class BaseMC(ABC):
             with an attached calculator, or a pre-initialized
             :class:`~clease.montecarlo.mc_evaluator.MCEvaluator`
             object.
+        temp (float): Temperature of Monte Carlo simulation in Kelvin
     """
 
-    def __init__(
-        self,
-        system: Union[Atoms, MCEvaluator],
-    ):
-
+    def __init__(self, system: Union[Atoms, MCEvaluator], temp: float):
         self.evaluator = system
+        self.temperature = temp
 
     @property
     def evaluator(self) -> MCEvaluator:
@@ -51,3 +49,38 @@ class BaseMC(ABC):
         :type: :class:`ase.atoms.Atoms`
         """
         return self.evaluator.atoms
+
+    def _get_temperature(self) -> float:
+        """Retrieve the temperature of the MC."""
+        return self._temperature
+
+    def _set_temperature(self, value: float) -> None:
+        """Set the internal temperature."""
+        self._temperature = value  # pylint: disable=attribute-defined-outside-init
+        self._on_temp_change()
+
+    def _on_temp_change(self) -> None:
+        """An MC object may choose to make adjustments after a temperature change,
+        by subclassing and adjusting this method."""
+
+    # Ensure that changing the temperature triggers the correct responses.
+    temperature = property(
+        _get_temperature,
+        _set_temperature,
+        doc="""
+        Property for getting and setting the temperature of the MC object.
+
+        :type: float
+        """,
+    )
+    # For backwards compatibility, alias the "temperature" as "T" as well
+    T = property(
+        _get_temperature,
+        _set_temperature,
+        doc="""
+                 Alias for the temperature variable.
+                 This variable name is deprecated in favor of :py:attr:`~temperature`.
+
+                 :type: float
+                 """,
+    )